+++ R/sjPlotKfoldCV.R
@@ -10,25 +10,28 @@ utils::globalVariables(c("train", "model", "test", ".response", "sse", "sst"))
 #'
 #' @param data A data frame, used to split the data into \code{k} trainig-test-pairs.
 #' @param formula A model formula, used to fit linear models (\code{\link[stats]{lm}})
-#'          over all \code{k} training data sets.
+#'          over all \code{k} training data sets. Use \code{fit} to specify a 
+#'          fitted model (also other models than linear models), which will be used
+#'          to compute cross validation. If \code{fit} is not missing, \code{formula}
+#'          will be ignored.
 #' @param k Number of folds.
+#' @param fit Model object, which will be used to compute cross validation. If 
+#'          \code{fit} is not missing, \code{formula} will be ignored. Currently,
+#'          only linear, poisson and negative binomial regression models are supported.
 #'
 #' @details This function, first, generates \code{k} cross-validated test-training
 #'            pairs (using the \code{\link[modelr]{crossv_kfold}}-function) and
-#'            fits the same linear model, specified in the \code{formula}-argument,
-#'            over all training data sets. \cr \cr
-#'            Then, the quality measures Root Mean Squared Error (\emph{rmse}),
-#'            R-squared and Mean Squared Error (\emph{mse}) for each model are 
-#'            calculated. The average for each of these measures is returned 
-#'            from this function. \cr \cr
-#'            Finally, the test data is used to predict the outcome from all
+#'            fits the same model, specified in the \code{formula}- or \code{fit}-
+#'            argument, over all training data sets. \cr \cr
+#'            Then, the test data is used to predict the outcome from all
 #'            models that have been fit on the training data, and the residuals
 #'            from all test data is plotted against the observed values (outcome)
-#'            from the test data. This plot can be used to validate the model
+#'            from the test data (note: for poisson or negative binomial models, the
+#'            deviance residuals are calculated). This plot can be used to validate the model
 #'            and see, whether it over- (residuals > 0) or underestimates 
 #'            (residuals < 0) the model's outcome.
 #'
-#' @note Currently, only linear models are supported.
+#' @note Currently, only linear, poisson and negative binomial regression models are supported.
 #'
 #' @examples 
 #' library(sjmisc)
@@ -36,6 +39,14 @@ utils::globalVariables(c("train", "model", "test", ".response", "sse", "sst"))
 #' 
 #' sjp.kfold_cv(efc, neg_c_7 ~ e42dep + c172code + c12hour)
 #' sjp.kfold_cv(mtcars, mpg ~.)
+#'
+#' # for poisson models. need to fit a model and use 'fit'-argument
+#' fit <- glm(tot_sc_e ~ neg_c_7 + c172code, data = efc, family = poisson)
+#' sjp.kfold_cv(efc, fit = fit)
+#' 
+#' # and for negative binomial models
+#' fit <- MASS::glm.nb(tot_sc_e ~ neg_c_7 + c172code, data = efc)
+#' sjp.kfold_cv(efc, fit = fit)
 #'
 #' @import ggplot2
 #' @importFrom tibble is.tibble as_tibble
@@ -46,46 +57,79 @@ utils::globalVariables(c("train", "model", "test", ".response", "sse", "sst"))
 #' @importFrom broom augment
 #' @importFrom tidyr unnest
 #' @importFrom graphics plot
-#' @importFrom stats as.formula
+#' @importFrom stats as.formula formula family poisson glm lm
+#' @importFrom sjstats resp_val
+#' @importFrom purrr map
+#' @importFrom MASS glm.nb
 #' @export
-sjp.kfold_cv <- function(data, formula, k = 5) {
+sjp.kfold_cv <- function(data, formula, k = 5, fit) {
   # make sure that data is a tibble
   if (!tibble::is.tibble(data)) data <- tibble::as_tibble(data)
   
-  # make sure we have a formula
-  if (!inherits(formula, "formula")) formula <- stats::as.formula(formula)
+  # check if a formula was passed as argument...
+  if (!missing(formula)) {
+    # make sure we have a formula
+    if (!inherits(formula, "formula")) formula <- stats::as.formula(formula)
+    # reset fam
+    fam <- NULL
+  } else if (!missing(fit)) {
+    # ... or a fitted model
+    formula <- stats::formula(fit)
+    
+    # get model family for glm
+    if (inherits(fit, "glm"))
+      fam <- stats::family(fit)
+    else
+      fam <- NULL
+  } else {
+    stop("Either `formula` or `fit` must be supplied.", call. = F)
+  }
   
   # get name of response variable and get variable label, if
   # there is any... used for labelling plot axis
   resp <- formula[[2]]
   resp.name <- sjmisc::get_label(data[[deparse(resp)]], def.value = deparse(resp))
+
+  # check if fit parameter was specified, and we have a model family
+  if (!is.null(fam)) {
+    # for poisson models, show deviance residuals
+    if (fam$family == "poisson") {
+      # create cross-validated test-training pairs, run poisson-model on each
+      # pair, get deviance residuals and response value
+      res <- modelr::crossv_kfold(data, k = k) %>% 
+        dplyr::mutate(model = purrr::map(train, ~ stats::glm(formula, data = .x, family = stats::poisson(link = "log")))) %>% 
+        dplyr::mutate(residuals = purrr::map(model, ~ stats::residuals(.x, "deviance"))) %>% 
+        dplyr::mutate(.response = purrr::map(model, ~ sjstats::resp_val(.x)))
+    # for negative binomial models, show deviance residuals
+    } else if (inherits(fit, "negbin")) {
+      # create cross-validated test-training pairs, run poisson-model on each
+      # pair, get deviance residuals and response value
+      res <- modelr::crossv_kfold(data, k = k) %>% 
+        dplyr::mutate(model = purrr::map(train, ~ MASS::glm.nb(formula, data = .))) %>% 
+        dplyr::mutate(residuals = purrr::map(model, ~ stats::residuals(.x, "deviance"))) %>% 
+        dplyr::mutate(.response = purrr::map(model, ~ sjstats::resp_val(.x)))
+    }
     
-  # create cross-validated test-training pairs, run linear model on each
-  # pair, get predicted values and quality measures for models fitted on the
-  # train data
-  res <- modelr::crossv_kfold(data, k = k) %>% 
-    dplyr::mutate(model = purrr::map(train, ~ lm(formula, data = .))) %>% 
-    dplyr::mutate(predicted = purrr::map2(model, test, ~ broom::augment(.x, newdata = .y))) %>% 
-    tidyr::unnest(predicted)
-
-  # make sure that response vector has an identifiably name
-  colnames(res)[which(colnames(res) == deparse(resp))] <- ".response"
-
-  # compute residuals for each k-fold model
-  res <- res %>% 
-    dplyr::mutate(residuals = .response - .fitted)
-  
-  # compute model quality measures r2 and rmse
-  gof <- res %>% 
-    dplyr::group_by_(".id") %>% 
-    dplyr::summarise(
-      sst = sum((.response - mean(.response, na.rm = T)) ^ 2, na.rm = T), # Sum of Squares Total
-      sse = sum(residuals ^ 2, na.rm = T),                                # Sum of Squares Residual/Error
-      rsquared = 1 - sse / sst,                                           # Proportion of variance accounted for
-      rmse = sqrt(mean(residuals ^ 2, na.rm = T))
-    ) %>% 
-    dplyr::ungroup()
-  
+    # unnest residuals and response values
+    res <- res %>% tidyr::unnest(residuals, .response)
+    
+  } else {
+    # create cross-validated test-training pairs, run linear model on each
+    # pair, get predicted values and quality measures for models fitted on the
+    # train data
+    res <- modelr::crossv_kfold(data, k = k) %>% 
+      dplyr::mutate(model = purrr::map(train, ~ stats::lm(formula, data = .))) %>% 
+      dplyr::mutate(predicted = purrr::map2(model, test, ~ broom::augment(.x, newdata = .y))) %>% 
+      tidyr::unnest(predicted)
+    
+    # make sure that response vector has an identifiably name
+    colnames(res)[which(colnames(res) == deparse(resp))] <- ".response"
+    
+    # compute residuals for each k-fold model
+    res <- res %>% 
+      dplyr::mutate(residuals = .response - .fitted)
+  }
+      
   # plot response against residuals, to see where our model over- or
   # underestimates the outcome
   p <- ggplot(data = res, aes_string(x = ".response", y = "residuals")) +
@@ -94,18 +138,11 @@ sjp.kfold_cv <- function(data, formula, k = 5) {
     stat_smooth(method = "loess") +
     theme_minimal() +
     labs(y = "Residuals", x = resp.name)
+  
   # plot it
-  graphics::plot(p)
-
-  # ouput of quality measures to console  
-  cat("Quality Measures of k-Fold Cross-Validation\n")
-  cat(sprintf("   R-squared: %.3f\n", mean(gof$rsquared)))
-  cat(sprintf("        RMSE: %.3f\n", mean(gof$rmse)))
+  suppressWarnings(graphics::plot(p))
 
   # return plot and quality measures
   invisible(structure(class = "sjp.kfold_cv",
-                      list(plot = p,
-                           rsquared = mean(gof$rsquared),
-                           rmse = mean(gof$rmse),
-                           mse = mean(gof$sse))))
+                      list(plot = p)))
 }